requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
anthropic==0.18.1
langgraph==0.3.1
browser-use==0.1.0
pydantic==2.6.1
python-multipart==0.0.9